+++ internal/decryption/key_provider_opts.go
@@ -1,0 +1,48 @@
+// Copyright 2022 MongoDB Inc
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package decryption
+
+type KeyProviderOpts struct {
+	Local KeyProviderLocalOpts
+	KMIP  KeyProviderKMIPOpts
+	AWS   KeyProviderAWSOpts
+	GCP   KeyProviderGCPOpts
+	Azure KeyProviderAzureOpts
+}
+
+type KeyProviderLocalOpts struct {
+	KeyFileName string
+}
+
+type KeyProviderKMIPOpts struct {
+	ServerCAFileName          string
+	ClientCertificateFileName string
+}
+
+type KeyProviderAWSOpts struct {
+	AccessKey       string
+	SecretAccessKey string
+	SessionToken    string
+}
+
+type KeyProviderGCPOpts struct {
+	ServiceAccountKey string
+}
+
+type KeyProviderAzureOpts struct {
+	ClientID string
+	TenantID string
+	Secret   string
+}
